requirements.txt
python-dotenv==1.0.1
ipywidgets==8.1.3
notebook==7.2.1
pandas==2.0.3
scikit-learn==1.3.2
google-generativeai==0.7.0
elasticsearch==8.14.0
ollama==0.2.1
streamlit==1.36.0
black[jupyter]==24.4.2